refactor(logger,utils): move normalizeError from utils into logger package
normalizeError is only used internally by sanitizeError in logger, so
keeping it in utils created an unnecessary cross-package dependency.
Moving it removes @unique-ag/utils from logger's runtime dependencies.
