+++ packages/logger/package.json
@@ -14,7 +14,6 @@
     "check-types": "tsc --noEmit"
   },
   "dependencies": {
-    "@unique-ag/utils": "workspace:*",
     "remeda": "2.32.0",
     "serialize-error-cjs": "0.2.0"
   },
