+++ packages/utils/src/index.ts
@@ -5,7 +5,6 @@ export {
   getHttpStatusCodeClass,
   getSlowRequestDurationBucket,
 } from './metrics';
-export { normalizeError } from './normalize-error';
 export {
   type BatchProcessorOptions,
   processInBatches,
